New folder/RealEstateHub/RealEstateHub/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0
django-allauth==65.13.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==40.1.0
idna==3.11
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.6.2
